backend/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
mongomock==4.1.2
packaging==23.1
pluggy==1.2.0
pymongo==4.5.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
sentinels==1.0.0
tomli==2.0.1
urllib3==2.0.4
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
